# Conflicts:
#	app/utilities/TestDatabaseManager.java
#	app/views/users/destination/indexDestination.scala.html
#	test/controllers/DestinationControllerTest.java
#	test/resources/test_photos/user_2/profilethumbnail.png

+++ app/controllers/DestinationController.java
@@ -3,26 +3,20 @@ package controllers;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import factories.DestinationFactory;
-
 import formdata.DestinationFormData;
-import formdata.UpdateUserFormData;
 import models.*;
-
-
 import play.data.DynamicForm;
 import play.data.Form;
 import play.data.FormFactory;
 import play.libs.Json;
-import play.mvc.BodyParser;
 import play.mvc.Controller;
 import play.mvc.Http;
 import play.mvc.Result;
+import views.html.users.destination.*;
+
 import javax.inject.Inject;
 import java.io.File;
 import java.util.*;
-
-
-import views.html.users.destination.*;
 
 public class DestinationController extends Controller {
 
@@ -481,9 +475,15 @@ public class DestinationController extends Controller {
             if (destination != null) {
                 if (destination.isUserOwner(user.userid) || user.userIsAdmin()) {
                     if(destination.visits.isEmpty()) {
-                        destination.delete();
-                        return redirect(routes.DestinationController.indexDestination());
-                    } else {
+                        List<TreasureHunt> treasureHunts = TreasureHunt.find.query().where().eq("destination", destination).findList();
+                        if (treasureHunts.isEmpty()) {
+                            destination.delete();
+                            return redirect(routes.DestinationController.indexDestination());
+                        } else {
+                            return preconditionRequired("You cannot delete destinations while they are being used by the treasure hunts.");
+                        }
+                    }
+                    else{
                         return preconditionRequired("You cannot delete destinations while you're using them for your trips. Delete them from your trip first!");
                     }
                 } else {
@@ -609,8 +609,8 @@ public class DestinationController extends Controller {
      * Returns a json list of photos associated to a destination given by a destination id
      *
      * @param request the HTTP request
-     * @param destId  the destination id
-     * @return a json list of traveller types associated to the destination
+     * @param destId the destination id
+     * @return a json list of photos associated to the destination
      */
     public Result getPhotos(Http.Request request, Integer destId) {
         User user = User.getCurrentUser(request);
@@ -720,8 +720,19 @@ public class DestinationController extends Controller {
     }
 
     /**
-     * Gets JSON of all visible (public + the logged in users private photos)
-     * Destinations avaliable to the user.
+     * Gets List of all destinations visible to the given user
+     * @param userId the user to look for private destinations of
+     * @return The list of all public destinations and all private destinations that the user can see
+     */
+    private List<Destination> getVisibleDestinations(int userId) {
+        DestinationFactory destinationFactory = new DestinationFactory();
+
+        return destinationFactory.getAllVisibleDestinations(userId);
+    }
+
+    /**
+     * Gets JSON of all visible (public + the logged in users private)
+     * Destinations available to the user.
      *
      * @param request the HTTP request
      * @return a Result object containing the destinations JSON in it's body
@@ -731,18 +742,46 @@ public class DestinationController extends Controller {
         if (user != null) {
             int userId = user.getUserid();
 
-            DestinationFactory destinationFactory = new DestinationFactory();
-
-            List<Destination> publicDestinations;
-            List<Destination> privateDestinations;
-            publicDestinations = destinationFactory.getPublicDestinations();
-            privateDestinations = destinationFactory.getUsersPrivateDestinations(userId);
-
-            List<Destination> allVisibleDestination = new ArrayList<Destination>();
-            allVisibleDestination.addAll(publicDestinations);
-            allVisibleDestination.addAll(privateDestinations);
+            List<Destination> allVisibleDestination = getVisibleDestinations(userId);
 
             return ok(Json.toJson(allVisibleDestination));
+        } else {
+            return unauthorized("Oops, you are not logged in");
+        }
+    }
+
+    /**
+     * Adds a photo with a photo id to a destination with a destination id.
+     * @param request the HTTP request
+     * @param photoId the photoId of the photo to e added
+     * @param destId the destination that the photo should be linked to
+     * @return success if the linking was successful, not found if destination or photo not found, unauthorized otherwise.
+     */
+    public Result addPhotoToDestination(Http.Request request, Integer photoId, Integer destId){
+        User user = User.getCurrentUser(request);
+        if(user != null) {
+            UserPhoto photo = UserPhoto.find.byId(photoId);
+            Destination destination = Destination.find.byId(destId);
+            if(destination != null && photo != null) {
+                if (photo.getUser().getUserid() == user.getUserid()) {
+                    //add checks for private destinations here once destinations have been merged in.
+                    //You can only link a photo to a private destination if you own the private destination.
+                    if(!photo.getDestinations().contains(destination)) {
+                        photo.addDestination(destination);
+                        photo.update();
+                        System.out.println("SUCCESS!");
+                        return redirect(routes.DestinationController.indexDestination());
+                    }
+                    else{
+                        return badRequest("You have already linked the photo to this destination.");
+                    }
+                } else {
+                    return unauthorized("Oops, this is not your photo!");
+                }
+            }
+            else{
+                return notFound();
+            }
         } else {
             return unauthorized("Oops, you are not logged in");
         }
